logic previously added to redis-storage moved to sai-session-storage
