+++ src/auth.ts
@@ -2,8 +2,7 @@ import { randomUUID } from "crypto";
 import express, { Response, Request } from "express";
 import { getSessionFromStorage, Session } from "@inrupt/solid-client-authn-node";
 
-import { RedisStorage } from "./redis-storage";
-import { uuid2clientId } from "./sai-session-storage";
+import storage, { uuid2clientId } from "./sai-session-storage";
 
 export const redirectUrl = `${process.env.BASE_URL}/auth/handleLoginRedirect`
 
@@ -21,7 +20,7 @@ router.post("/login", async (req: Request, res: Response) => {
   }
 
   const oidcSession = new Session({
-    storage: RedisStorage.instance
+    storage: storage.oidcStorage
   });
 
   req.session["sessionId"] = oidcSession.info.sessionId;
@@ -46,7 +45,7 @@ router.get("/handleLoginRedirect", async (req: Request, res: Response) => {
     return;
   }
 
-  const oidcSession = await getSessionFromStorage(req.session["sessionId"], RedisStorage.instance);
+  const oidcSession = await getSessionFromStorage(req.session["sessionId"], storage.oidcStorage);
 
   if (!oidcSession) {
     res.redirect(401, process.env.BASE_URL + "/login");
@@ -56,7 +55,7 @@ router.get("/handleLoginRedirect", async (req: Request, res: Response) => {
   await oidcSession.handleIncomingRedirect(process.env.BASE_URL + "/auth" + req.url);
 
   if (oidcSession.info.isLoggedIn && oidcSession.info.webId) {
-    await RedisStorage.instance.rename(req.session["sessionId"], oidcSession.info.webId)
+    await storage.changeKey(req.session["sessionId"], oidcSession.info.webId)
     delete req.session["sessionId"]
     req.session["webId"] = oidcSession.info.webId;
     res.redirect(200, process.env.BASE_URL + "/dashboard");
